+++ CHANGES.md
@@ -1,4 +1,9 @@
-### 3.1.0 - XXXX-XX-XX
+### 3.1.0 - 2022-02-22
+
+- Fix loky.cpu_count() to properly detect the number of allowed CPUs based on
+  the /sys/fs/cgroup/cpu.max file on newest Linux versions with cgroup v2.
+  Fall-back to the /sys/fs/cgroup/cpu/cpu.cfs_quota_us file to keep on
+  supporting Linux versions that use cgroup v1 (#355 and #358).
 
 - Fix an exception that could be raised in an auxiliary thread when
   garbage collecting an executor instance when shutting down the
@@ -6,7 +11,10 @@
 
 - Make `shutdown(kill_workers=True)` consistently use the SIGKILL
   signal on POSIX. Previously a mix of SIGKILL and SIGTERM was issued
-  and could deadlock the shutdown process (#348).
+  and could deadlock the shutdown process (#348 and #357).
+
+- Big code clean-up to drop support for older Python versions.
+  Python 3.7 or later is now required. (#304)
 
 ### 3.0.0 - 2021-09-10